Adding wasm to d_std
Former-commit-id: 5d95c908d592b23fabd407412ac403a31ba92b5d
Former-commit-id: 07fadd7a0140d393eb75c8316dbe707cc98ed677 [formerly 6ed167fff6ffb0474d5c6457e0804bee6aca2b49]
Former-commit-id: 335d451f1ff34170a76b34114a62bd10614188a6

+++ modules/util/source/hip/util/string.d
@@ -632,7 +632,7 @@ string getNumericEnding(string s)
 {
     if(!s)
         return "";
-    long i = cast(long)s.length - 1;
+    ptrdiff_t i = cast(ptrdiff_t)s.length - 1;
     while(i >= 0)
     {
         if(!isNumeric(s[i]))
